// src/types/catalog/products/addProductRequest.ts
export interface IAddProductRequestRoot {
  id?: number;
  userId?: number;
  name?: string;
  type?: string;
  description?: string;
  supplyPrice?: number;
  maxRetailPrice?: number;
  retailPrice?: number;
  sku?: string;
  barcode?: string;
  barcodeStrategy?: string;
  quantity?: number;
  uom?: number;
  supplierId?: number;
  categoryId?: number;
  brandId?: number;
  tags?: string;
  height?: number;
  width?: number;
  length?: number;
  weight?: number;
  strategy?: string;
  expiryDays?: number;
  trackSerialNumbers?: boolean;
  trackExpiryDates?: boolean;
  syncSupplyPrice?: boolean;
  image?: string[];
  fileUrl?: string;
}